requirements.txt
#langchain core
langchain
langchain-core

#openai Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini Integration(PALM)
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface_hub

python-dotenv

numpy
scikit-learn
pydantic
 

streamlit
